feat(lms): add DTOs for module tree and refactor module content
`ModuleTreeDto` and `ModuleTreeSectionDto` provide a hierarchical
representation of modules and their sections. `BasicModuleItemDto`
extracts common module content properties for reuse across different
DTOs, including `ModuleTreeDto` and `PaginatedModuleContentDto`.
